+++ CHANGELOG.md
@@ -19,3 +19,4 @@ used to track changes in this project.
 ### Changed
 - Migrate to UUIDs for primary keys
 - LoginWithTokenQuery to LoginWithTokenCommand
+- Moved raising OrderShippedEvent from OrderService::addShipment() to Order::addShipment()